Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    int number;
-    scanf("%d",&number);
-    printf("You entered: ",number);
+    float number;
+    scanf("%f",&number);
+    printf("You entered: ", number);
     return 0;
 }
